backend/celery_config.py
from celery import Celery
import os
from dotenv import load_dotenv

load_dotenv()

celery_app = Celery(
    "rag_worker",
    broker=os.getenv("CELERY_BROKER_URL", "redis://localhost:6379/0"),
    backend=os.getenv("CELERY_RESULT_BACKEND", "redis://localhost:6379/0"),
    include=["workers.pdf_worker"]
)

celery_app.conf.update(
    task_serializer="json",
    result_serializer="json",
    accept_content=["json"],
    task_track_started=True,
    worker_pool="solo",  # Required for Windows
)